@@ -1,8 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFrameworks>net6.0;net7.0</TargetFrameworks>
-		<Copyright>Copyright 算神工作室 2022.</Copyright>
+		<TargetFrameworks>net7.0</TargetFrameworks>
+		<Copyright>Copyright 算神工作室 2023.</Copyright>
 		<Authors>算神</Authors>
 		<PackageId>JT808.Protocol.Extensions.Streamax</PackageId>
 		<Product>JT808.Protocol.Extensions.Streamax</Product>
@@ -10,7 +10,7 @@
 		<PackageReleaseNotes>基于JT808协议锐明公交业务扩展部分</PackageReleaseNotes>
 		<RepositoryUrl>https://github.com/lishewen/Streamax.Protocol</RepositoryUrl>
 		<PackageProjectUrl>https://github.com/lishewen/Streamax.Protocol</PackageProjectUrl>
-		<Version>0.4.3</Version>
+		<Version>0.4.4</Version>
 		<licenseUrl>https://github.com/lishewen/Streamax.Protocol/blob/master/LICENSE</licenseUrl>
 		<license>https://github.com/lishewen/Streamax.Protocol/blob/master/LICENSE</license>
 		<DocumentationFile>JT808.Protocol.Extensions.Streamax.xml</DocumentationFile>
@@ -20,7 +20,7 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="JT808" Version="2.5.1" />
+		<PackageReference Include="JT808" Version="2.6.0" />
 		<None Include="..\readme.md" Pack="true" PackagePath="\" />
 		<None Include="..\LICENSE" Pack="true" PackagePath="\" />
 	</ItemGroup>
