@@ -12,17 +12,17 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
     /// <summary>
     /// 视频相关报警
     /// </summary>
-    public class JT808_0x0200_0x14 : JT808_0x0200_CustomBodyBase, IJT808MessagePackFormatter<JT808_0x0200_0x14>, IJT808Analyze
+    public class JT808_0x0200_0x14 : JT808MessagePackFormatter<JT808_0x0200_0x14>, JT808_0x0200_CustomBodyBase, IJT808Analyze
     {
-        public override byte AttachInfoId { get; set; } = 0x14;
-        public override byte AttachInfoLength { get; set; } = 4;
+        public byte AttachInfoId { get; set; } = 0x14;
+        public byte AttachInfoLength { get; set; } = 4;
         /// <summary>
         /// 视频报警标志位
         /// </summary>
         public uint AlarmTag { get; set; }
         public void Analyze(ref JT808MessagePackReader reader, Utf8JsonWriter writer, IJT808Config config)
         {
-            JT808_0x0200_0x14 value = new JT808_0x0200_0x14();
+            JT808_0x0200_0x14 value = new();
             value.AttachInfoId = reader.ReadByte();
             writer.WriteNumber($"[{value.AttachInfoId.ReadNumber()}]附加信息Id", value.AttachInfoId);
             value.AttachInfoLength = reader.ReadByte();
@@ -39,16 +39,16 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
             writer.WriteEndArray();
         }
 
-        public JT808_0x0200_0x14 Deserialize(ref JT808MessagePackReader reader, IJT808Config config)
+        public override JT808_0x0200_0x14 Deserialize(ref JT808MessagePackReader reader, IJT808Config config)
         {
-            JT808_0x0200_0x14 value = new JT808_0x0200_0x14();
+            JT808_0x0200_0x14 value = new();
             value.AttachInfoId = reader.ReadByte();
             value.AttachInfoLength = reader.ReadByte();
             value.AlarmTag = reader.ReadUInt32();
             return value;
         }
 
-        public void Serialize(ref JT808MessagePackWriter writer, JT808_0x0200_0x14 value, IJT808Config config)
+        public override void Serialize(ref JT808MessagePackWriter writer, JT808_0x0200_0x14 value, IJT808Config config)
         {
             writer.WriteByte(value.AttachInfoId);
             writer.WriteByte(value.AttachInfoLength);
