@@ -11,11 +11,11 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
     /// <summary>
     /// 运营登记
     /// </summary>
-    public class JT808_0x0B01 : JT808Bodies, IJT808MessagePackFormatter<JT808_0x0B01>, IJT808Analyze
+    public class JT808_0x0B01 : JT808MessagePackFormatter<JT808_0x0B01>, JT808Bodies, IJT808Analyze
     {
-        public override ushort MsgId => 0x0B01;
+        public ushort MsgId => 0x0B01;
 
-        public override string Description => "运营登记";
+        public string Description => "运营登记";
         /// <summary>
         /// 线路编号
         /// </summary>
@@ -26,7 +26,7 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
         public string WorkerId { get; set; }
         public void Analyze(ref JT808MessagePackReader reader, Utf8JsonWriter writer, IJT808Config config)
         {
-            JT808_0x0B01 value = new JT808_0x0B01();
+            JT808_0x0B01 value = new();
             value.GprsId = reader.ReadUInt32();
             writer.WriteNumber($"[{value.GprsId.ReadNumber()}]线路编号", value.GprsId);
             var length = reader.ReadCurrentRemainContentLength();
@@ -35,16 +35,16 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
             writer.WriteString($"[{virtualHex.ToArray().ToHexString()}]员工编号", value.WorkerId);
         }
 
-        public JT808_0x0B01 Deserialize(ref JT808MessagePackReader reader, IJT808Config config)
+        public override JT808_0x0B01 Deserialize(ref JT808MessagePackReader reader, IJT808Config config)
         {
-            JT808_0x0B01 value = new JT808_0x0B01();
+            JT808_0x0B01 value = new();
             value.GprsId = reader.ReadUInt32();
             var length = reader.ReadCurrentRemainContentLength();
             value.WorkerId = reader.ReadString(length);
             return value;
         }
 
-        public void Serialize(ref JT808MessagePackWriter writer, JT808_0x0B01 value, IJT808Config config)
+        public override void Serialize(ref JT808MessagePackWriter writer, JT808_0x0B01 value, IJT808Config config)
         {
             writer.WriteUInt32(value.GprsId);
             writer.WriteString(value.WorkerId);
