@@ -11,11 +11,11 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
     /// <summary>
     /// 行车计划请求
     /// </summary>
-    public class JT808_0x0B07 : JT808Bodies, IJT808MessagePackFormatter<JT808_0x0B07>, IJT808Analyze
+    public class JT808_0x0B07 : JT808MessagePackFormatter<JT808_0x0B07>, JT808Bodies, IJT808Analyze
     {
-        public override ushort MsgId => 0x0B07;
+        public ushort MsgId => 0x0B07;
 
-        public override string Description => "行车计划请求";
+        public string Description => "行车计划请求";
         /// <summary>
         /// 营运日期
         /// BCD[3]
@@ -28,7 +28,7 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
         public string WorkerId { get; set; }
         public void Analyze(ref JT808MessagePackReader reader, Utf8JsonWriter writer, IJT808Config config)
         {
-            JT808_0x0B07 value = new JT808_0x0B07();
+            JT808_0x0B07 value = new();
             value.WorkDate = reader.ReadDateTime_YYMMDD();
             writer.WriteString($"[{value.WorkDate:yyMMdd}]营运日期", value.WorkDate.ToString("yy-MM-dd"));
             var length = reader.ReadCurrentRemainContentLength();
@@ -37,16 +37,16 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
             writer.WriteString($"[{virtualHex.ToArray().ToHexString()}]员工编号", value.WorkerId);
         }
 
-        public JT808_0x0B07 Deserialize(ref JT808MessagePackReader reader, IJT808Config config)
+        public override JT808_0x0B07 Deserialize(ref JT808MessagePackReader reader, IJT808Config config)
         {
-            JT808_0x0B07 value = new JT808_0x0B07();
+            JT808_0x0B07 value = new();
             value.WorkDate = reader.ReadDateTime_YYMMDD();
             var length = reader.ReadCurrentRemainContentLength();
             value.WorkerId = reader.ReadString(length);
             return value;
         }
 
-        public void Serialize(ref JT808MessagePackWriter writer, JT808_0x0B07 value, IJT808Config config)
+        public override void Serialize(ref JT808MessagePackWriter writer, JT808_0x0B07 value, IJT808Config config)
         {
             writer.WriteDateTime_YYMMDD(value.WorkDate);
             writer.WriteString(value.WorkerId);
