@@ -12,11 +12,11 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
     /// <summary>
     /// 业务请求
     /// </summary>
-    public class JT808_0x0B09 : JT808Bodies, IJT808MessagePackFormatter<JT808_0x0B09>, IJT808Analyze
+    public class JT808_0x0B09 : JT808MessagePackFormatter<JT808_0x0B09>, JT808Bodies, IJT808Analyze
     {
-        public override ushort MsgId => 0x0B09;
+        public ushort MsgId => 0x0B09;
 
-        public override string Description => "业务请求";
+        public string Description => "业务请求";
         /// <summary>
         /// 线路编号
         /// </summary>
@@ -37,7 +37,7 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
         public DateTime Time { get; set; }
         public void Analyze(ref JT808MessagePackReader reader, Utf8JsonWriter writer, IJT808Config config)
         {
-            JT808_0x0B09 value = new JT808_0x0B09();
+            JT808_0x0B09 value = new();
             value.GprsId = reader.ReadUInt32();
             writer.WriteNumber($"[{value.GprsId.ReadNumber()}]线路编号", value.GprsId);
             var length = reader.ReadCurrentRemainContentLength() - 7;
@@ -50,9 +50,9 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
             writer.WriteString($"[{value.Time:yyMMddHHmmss}]时间", value.Time.ToString("yyyy-MM-dd HH:mm:ss"));
         }
 
-        public JT808_0x0B09 Deserialize(ref JT808MessagePackReader reader, IJT808Config config)
+        public override JT808_0x0B09 Deserialize(ref JT808MessagePackReader reader, IJT808Config config)
         {
-            JT808_0x0B09 value = new JT808_0x0B09();
+            JT808_0x0B09 value = new();
             value.GprsId = reader.ReadUInt32();
             var length = reader.ReadCurrentRemainContentLength() - 7;
             value.WorkerId = reader.ReadString(length);
@@ -61,7 +61,7 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
             return value;
         }
 
-        public void Serialize(ref JT808MessagePackWriter writer, JT808_0x0B09 value, IJT808Config config)
+        public override void Serialize(ref JT808MessagePackWriter writer, JT808_0x0B09 value, IJT808Config config)
         {
             writer.WriteUInt32(value.GprsId);
             writer.WriteString(value.WorkerId);
