@@ -12,11 +12,11 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
     /// <summary>
     /// 设备故障上报
     /// </summary>
-    public class JT808_0x0B0B : JT808Bodies, IJT808MessagePackFormatter<JT808_0x0B0B>, IJT808Analyze
+    public class JT808_0x0B0B : JT808MessagePackFormatter<JT808_0x0B0B>, JT808Bodies, IJT808Analyze
     {
-        public override ushort MsgId => 0x0B0B;
+        public ushort MsgId => 0x0B0B;
 
-        public override string Description => "设备故障上报";
+        public string Description => "设备故障上报";
         /// <summary>
         /// 设备类型
         /// </summary>
@@ -41,7 +41,7 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
         public DateTime Time { get; set; }
         public void Analyze(ref JT808MessagePackReader reader, Utf8JsonWriter writer, IJT808Config config)
         {
-            JT808_0x0B0B value = new JT808_0x0B0B();
+            JT808_0x0B0B value = new();
             value.DeviceType = reader.ReadByte();
             writer.WriteNumber($"[{value.DeviceType.ReadNumber()}]设备类型-{Enum.GetName(typeof(DeviceType), value.DeviceType)}", value.DeviceType);
             value.DeviceAddress = reader.ReadByte();
@@ -56,9 +56,9 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
             writer.WriteString($"[{value.Time:yyMMddHHmmss}]时间", value.Time.ToString("yyyy-MM-dd HH:mm:ss"));
         }
 
-        public JT808_0x0B0B Deserialize(ref JT808MessagePackReader reader, IJT808Config config)
+        public override JT808_0x0B0B Deserialize(ref JT808MessagePackReader reader, IJT808Config config)
         {
-            JT808_0x0B0B value = new JT808_0x0B0B();
+            JT808_0x0B0B value = new();
             value.DeviceType = reader.ReadByte();
             value.DeviceAddress = reader.ReadByte();
             var length = reader.ReadCurrentRemainContentLength() - 7;
@@ -68,7 +68,7 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
             return value;
         }
 
-        public void Serialize(ref JT808MessagePackWriter writer, JT808_0x0B0B value, IJT808Config config)
+        public override void Serialize(ref JT808MessagePackWriter writer, JT808_0x0B0B value, IJT808Config config)
         {
             writer.WriteByte(value.DeviceType);
             writer.WriteByte(value.DeviceAddress);
