@@ -12,11 +12,11 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
     /// <summary>
     /// 业务变更指令
     /// </summary>
-    public class JT808_0x8B02 : JT808Bodies, IJT808MessagePackFormatter<JT808_0x8B02>, IJT808Analyze
+    public class JT808_0x8B02 : JT808MessagePackFormatter<JT808_0x8B02>, JT808Bodies, IJT808Analyze
     {
-        public override ushort MsgId => 0x8B02;
+        public ushort MsgId => 0x8B02;
 
-        public override string Description => "业务变更指令";
+        public string Description => "业务变更指令";
         /// <summary>
         /// 线路编号
         /// </summary>
@@ -32,7 +32,7 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
         public string Additional { get; set; }
         public void Analyze(ref JT808MessagePackReader reader, Utf8JsonWriter writer, IJT808Config config)
         {
-            JT808_0x8B02 value = new JT808_0x8B02();
+            JT808_0x8B02 value = new();
             value.GprsId = reader.ReadUInt32();
             writer.WriteNumber($"[{value.GprsId.ReadNumber()}]线路编号", value.GprsId);
             value.TrafficType = reader.ReadByte();
@@ -42,16 +42,16 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
             writer.WriteString($"[{virtualHex.ToArray().ToHexString()}]附加内容", value.Additional);
         }
 
-        public JT808_0x8B02 Deserialize(ref JT808MessagePackReader reader, IJT808Config config)
+        public override JT808_0x8B02 Deserialize(ref JT808MessagePackReader reader, IJT808Config config)
         {
-            JT808_0x8B02 value = new JT808_0x8B02();
+            JT808_0x8B02 value = new();
             value.GprsId = reader.ReadUInt32();
             value.TrafficType = reader.ReadByte();
             value.Additional = reader.ReadStringEndChar0();
             return value;
         }
 
-        public void Serialize(ref JT808MessagePackWriter writer, JT808_0x8B02 value, IJT808Config config)
+        public override void Serialize(ref JT808MessagePackWriter writer, JT808_0x8B02 value, IJT808Config config)
         {
             writer.WriteUInt32(value.GprsId);
             writer.WriteByte(value.TrafficType);
