@@ -12,11 +12,11 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
     /// <summary>
     /// 业务请求应答
     /// </summary>
-    public class JT808_0x8B09 : JT808Bodies, IJT808MessagePackFormatter<JT808_0x8B09>, IJT808Analyze
+    public class JT808_0x8B09 : JT808MessagePackFormatter<JT808_0x8B09>, JT808Bodies, IJT808Analyze
     {
-        public override ushort MsgId => 0x8B09;
+        public ushort MsgId => 0x8B09;
 
-        public override string Description => "业务请求应答";
+        public string Description => "业务请求应答";
         /// <summary>
         /// 流水号
         /// </summary>
@@ -106,7 +106,7 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
         public string Additional { get; set; }
         public void Analyze(ref JT808MessagePackReader reader, Utf8JsonWriter writer, IJT808Config config)
         {
-            JT808_0x8B09 value = new JT808_0x8B09();
+            JT808_0x8B09 value = new();
             value.Serial = reader.ReadUInt16();
             writer.WriteNumber($"[{value.Serial.ReadNumber()}]流水号", value.Serial);
             value.Response = reader.ReadByte();
@@ -149,9 +149,9 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
             writer.WriteString($"[Length:{value.Additional.Length}]附加内容", value.Additional);
         }
 
-        public JT808_0x8B09 Deserialize(ref JT808MessagePackReader reader, IJT808Config config)
+        public override JT808_0x8B09 Deserialize(ref JT808MessagePackReader reader, IJT808Config config)
         {
-            JT808_0x8B09 value = new JT808_0x8B09();
+            JT808_0x8B09 value = new();
             value.Serial = reader.ReadUInt16();
             value.Response = reader.ReadByte();
             value.Time = reader.ReadDateTime_yyMMddHHmmss();
@@ -175,7 +175,7 @@ namespace JT808.Protocol.Extensions.Streamax.MessageBody
             return value;
         }
 
-        public void Serialize(ref JT808MessagePackWriter writer, JT808_0x8B09 value, IJT808Config config)
+        public override void Serialize(ref JT808MessagePackWriter writer, JT808_0x8B09 value, IJT808Config config)
         {
             writer.WriteUInt16(value.Serial);
             writer.WriteByte(value.Response);
